requirements.txt
cryptography
pyasn1
pyopenssl

# not mandatory:
netifaces